+++ extension/module/MOD-INF/module.properties
@@ -1,0 +1,3 @@
+name = biodec
+description = a short description of your extension goes here
+requires = core
