+++ extension/module/styles/project-injection.css
@@ -1,0 +1,5 @@
+/* CSS styles */
+
+:root {
+  --color: blue;
+}
