--- proxy.conf
@@ -1,11 +1,0 @@
-server {
-    listen 80;
-    server_name localhost;
-    access_log /var/log/nginx/localhost.access.log;
-
-    location PROXY_PREFIX/openrefine/ {
-        proxy_buffering off;
-        proxy_pass         http://127.0.0.1:3333/;
-        proxy_redirect     http://127.0.0.1:3333/ PROXY_PREFIX/openrefine/;
-    }
-}
